added seeds removed migration
