+++ server/.env.example
@@ -16,7 +16,7 @@ TYPEORM_PORT=5432
 TYPEORM_SYNCHRONIZE=true
 TYPEORM_LOGGING=true
 TYPEORM_ENTITIES=src/entity/**/*.ts
-TYPEORM_MIGRATIONS=src/migration/**/*.ts
+TYPEORM_MIGRATIONS=src/migration/**/*.ts,src/seeders/**/*.ts
 TYPEORM_MIGRATIONS_DIR=src/migration
 TYPEORM_ENTITIES_DIR=src/entity
